Collapse microblog helper layer
